Add `Dk::Pkg::Validate` task
This adds a `Validate` task that is setup for logic that will only
install packages via tasks if they haven't already been installed.

The `Validate` task will be run before any task that uses `Dk::Pkg`
logic to control when packages are installed. `Validate` will
ensure that a manifest path param has been set and will handle
parsing the manifest file. The parsed manifest results in an array
that will be stored in the params as the installed packages.
Eventually this will be used to check if a package has already been
installed and if so the logic to install it won't be run. This will
allow setting up package tasks that depend on one another but will
ensure they are only installed once. This will also allow
re-running a build of many package tasks and ensure it won't try
and reinstall packages.

This also commonizes the `Dk::Pkg` constants so they can be used
by `Validate` and the upcoming logic that controls whether a
package will be installed or not.