Add `InstallPkg` mixin
This adds an `InstallPkg` mixin for allowing tasks to install pkgs
using dk-pkg. This will allow tasks to group a set of running
tasks or commands as the steps to install a pkg. The mixin will
then track if that pkg has been installed and if so it won't try
to install it again. This will allow re-running groups of tasks
without worrying about messing up any existing pkgs that have
already been installed.

The `InstallPkg` mixin adds a private `install_pkg` helper method
which takes a pkg name and a block. The method will check if the
pkg name is in the installed pkgs param and if not, it will yield
to the block. After the block finishes the `install_pkg` method
will add the pkg name to the installed pkgs and run a command to
update the manifest file on disk. If the pkg name is included in
the install pkgs param then the block isn't run and it simply
logs that its already installed. This ensures that if the pkg is
encountered again (in the current or a subsequent run) then it
won't try to run the package's cmds again which could cause
problems.

The `InstallPkg` mixin also adds the `Validate` task as a before
callback which ensures that the current dk run has the manifest
file path configured and has parsed it. This ensures that the
`install_pkg` helper can work as expected.

This also breaks out the logic for reading and writing the manifest
into a `Manifest` module with serialize and deserialize methods.
The `Validate` task uses the deserialize method on the contents of
the manifest file. The `InstallPkg` mixin uses the serialize method
to write the manifest file after a pkg has been installed. This
centralizes the serialization logic and adds logic to clean the
pkgs when serializing/deserializing. This keeps the manifest file
clean and avoids edge cases that could cause errors.