Add `TestHelpers` to `InstallPkg`
This updates the `InstallPkg` mixin to provide test helpers. These
are needed to make testing tasks that use the mixin easier.

The test helpers provide a helper method for testing that a package
was installed using the `install_pkg` method and a method for
removing `InstallPkg` runs from the test runner. The test runner's
runs will include any task callbacks that `InstallPkg` adds and any
cmds that `install_pkg` runs. To avoid having to worry about these,
the `non_dk_install_pkg_runs` method will take a test runner and
remove runs that the install pkg logic caused. The runs array that
is returned can then be used normally to test which cmds or tasks
were run.

This also includes a minor test change to the
`Validate::TestHelpers` tests to ensure the test helpers mixin is
using `MuchPlugin`. Previously the class that the test helpers were mixed
into it was being tested for `MuchPlugin`.