Don't set manifest path ivar in `Validate::TestHelpers`
This updates `Validate::TestHelpers` to not set a manifest path
ivar and only conditionally set the manifest path param. This is
to allow apps using dk-pkg to set their own manifest path and use
that in the tests. This helps avoid confusion where the
`@dk_pkg_manifest_path` ivar is set but it doesn't match what is
in the params. This is also makes the test helpers more closely
match the live code in that the `Validate` task doesn't set a
manifest path.