Allow setting the manifest mode and owner
This adds a manifest mode and owner param that are used when the
manifest is created. If these are provided then when the manifest
is created it will `chmod` and `chown` the manifest file. If these
aren't provided then no `chmod` or `chown` cmds are run for the
manifest file. This allows more control over the manifest file and
its mode/owner.

The `Validate` task now uses the new params when it creates a
manifest file. It will now test if the manifest file exists and if
it doesn't, it will create it and optionally `chmod` and `chown`
it using the params. The `Validate` task will not modify the
mode or owner of the manifest file after it has been created.