Allow passing runs to `non_dk_install_pkg_runs`
This updates the `non_dk_install_pkg_runs` test helper to expect
it will be passed a collection of runs instead of always using the
runs from the runner that is passed to it. By taking a runs array
and returning a filtered runs array the test helper can be chained
together with other helpers that filter the runs array. This allows
a task's tests to filter out any runs that are not specific to it
which makes testing easier.

This also includes a ruby 2.3.1 cleanup so that the test suite
runs on the newer version of ruby.